refactor: extract offset grid into its own component
